+++ core/engine/src/lib.rs
@@ -89,6 +89,7 @@ pub mod class;
 pub mod context;
 pub mod environments;
 pub mod error;
+pub mod interop;
 pub mod job;
 pub mod module;
 pub mod native_function;
@@ -120,7 +121,8 @@ pub mod prelude {
         context::Context,
         error::{JsError, JsNativeError, JsNativeErrorKind},
         host_defined::HostDefined,
-        module::Module,
+        interop::{IntoJsFunctionCopied, UnsafeIntoJsFunction},
+        module::{IntoJsModule, Module},
         native_function::NativeFunction,
         object::{JsData, JsObject, NativeObject},
         script::Script,
