+++ core/interop/src/macros.rs
@@ -64,7 +64,8 @@
 /// ```
 /// # use boa_engine::{JsString, JsData, js_string};
 /// # use boa_gc::{Finalize, Trace};
-/// use boa_interop::{js_class, Ignore, JsClass};
+/// use boa_engine::interop::{Ignore, JsClass};
+/// use boa_interop::{js_class};
 ///
 /// #[derive(Clone, Trace, Finalize, JsData)]
 /// pub enum Animal {
@@ -245,7 +246,7 @@ macro_rules! js_class {
             ) -> $crate::boa_engine::JsResult<$class_name> {
                 let rest = args;
                 $(
-                    let ($ctor_arg, rest) : ($ctor_arg_ty, _) = $crate::TryFromJsArgument::try_from_js_argument(new_target, rest, context)?;
+                    let ($ctor_arg, rest) : ($ctor_arg_ty, _) = $crate::boa_engine::interop::TryFromJsArgument::try_from_js_argument(new_target, rest, context)?;
                 )*
 
                 $constructor_body
@@ -393,7 +394,8 @@ macro_rules! __get_set_decl {
 #[allow(clippy::too_many_lines)]
 fn js_class_test() {
     use crate::IntoJsFunctionCopied;
-    use crate::{js_class, loaders, JsClass};
+    use crate::{js_class, loaders};
+    use boa_engine::interop::JsClass;
     use boa_engine::property::Attribute;
     use boa_engine::{js_string, Context, JsData, JsError, JsResult, Module, Source};
     use boa_gc::{Finalize, Trace};
